Struct ethcore_io::IoChannel

pub struct IoChannel<Message> where
    Message: Send + Sync + 'static,  { /* fields omitted */ }

Allows sending messages into the event loop. All the IO handlers will get the message in the message callback.

Methods

impl<Message> IoChannel<Message> where
Message: Send + Sync + 'static,

pub fn send(&self, message: Message) -> Result<(), IoError>

Send a message through the channel

pub fn send_sync(&self, message: Message) -> Result<(), IoError>

Send a message through the channel and handle it synchronously

pub fn disconnected() -> IoChannel<Message>

Create a new channel disconnected from an event loop.