Struct ethcore_io::IoChannel

pub struct IoChannel<Message> where Message: Send + Clone { /* fields omitted */ }

Allows sending messages into the event loop. All the IO handlers will get the message in the message callback.

Methods

impl<Message> IoChannel<Message> where Message: Send + Clone + Sync + 'static

fn send(&self, message: Message) -> Result<(), IoError>

Send a message through the channel

fn send_sync(&self, message: Message) -> Result<(), IoError>

Send a message through the channel and handle it synchronously

fn send_io(&self, message: IoMessage<Message>) -> Result<(), IoError>

Send low level io message

fn disconnected() -> IoChannel<Message>

Create a new channel to connected to event loop.

fn to_handler(handler: Weak<IoHandler<Message>>) -> IoChannel<Message>

Create a new synchronous channel to a given handler.